refactor: rename WorkflowAgentProposalCoordinator to AgentProposalCoordinator
The coordinator handles both workflow and tool-use agent proposals,
not just workflow agents. Updated naming to reflect this:

- Rename file: WorkflowAgentProposalCoordinator.ts → AgentProposalCoordinator.ts
- Rename class: WorkflowAgentProposalCoordinatorImpl → AgentProposalCoordinatorImpl
- Rename provider methods: pendingWorkflowAgentProposals → pendingAgentProposals
- Rename handler methods to use AgentProposal instead of WorkflowAgentProposal
